Added The Sims
Script for the first game in the series. I've tried every wine version between 2.19-staging and 3.0 and the game crashed right before the loading screen no matter what I did. Seems to be some sort of a regression in wine. It also requires the "-skip_intro" line because the intro uses some really old codec wine does not support.