Update latest Wine versions automatically (#1153)
Provide functions to fetch the latest Wine stable/devel/staging version. These functions can be used as input to ".wineVersion()" in the QuickScripts.
As the implementation requires that the versions json is updated, a setup wizard is required when executing the functions (to show the download progress). Therefore, "QuickScript#_determineWineVersion()" is used to delay the function execution until the setup wizard has been created.
NOTE: every derived QuickScript must call this method such that "_wineVersion" is initialized.